Update to version 1.3.1
