+++ NEWS
@@ -1,5 +1,13 @@
 (P Versions are the Password support versions, we keep the same base "1.x.y" as the upstream version)
 
+1.3.1 - Released 11-Feb-2022
+--------------------------------
+- Fixed a potential crash in SSDP request parsing.
+- Fixed a configure script failure on some platforms.
+- Protect against DNS rebinding attacks.
+- Fix an socket leakage issue on some platforms.
+- Minor bug fixes.
+
 1.3.0.1P - Released 20-Aug-2021
 --------------------------------
 - Fixes in the 1.3.0 Migration (Thanks Vitor, PR #22)
