+++ buildroot/Config.in
@@ -1,0 +1,1 @@
+source "$BR2_EXTERNAL_READYMEDIA_PATH/readymedia/Config.in"
