+++ buildroot/external.desc
@@ -1,0 +1,2 @@
+name: READYMEDIA
+desc: ReadyMedia external tree
