+++ buildroot/external.mk
@@ -1,0 +1,1 @@
+include $(sort $(wildcard $(BR2_EXTERNAL_READYMEDIA_PATH)/*/*.mk))
